Add files via upload
This is a simplified version of an Example provided by Mathworks in the 2021a release. This is a two-model approach to achieve a high sampling rate with F28379D Launchpad from TI. One model is uploaded to Launchpad which is running with a sampling rate of 25000. The baud rate is set to 12e6 to accommodate this much data transfer. 
The other model is running in Simulink fetching data every 0.06 seconds.